+++ reflect/Cookie.go
@@ -1,6 +1,8 @@
 package reflect
 
 import (
+	"net/http"
+
 	lua "github.com/anaminus/gopher-lua"
 	"github.com/anaminus/rbxmk"
 	"github.com/anaminus/rbxmk/dump"
@@ -40,6 +42,31 @@ func Cookie() rbxmk.Reflector {
 						ReadOnly:    true,
 						Summary:     "Types/Cookie:Properties/Name/Summary",
 						Description: "Types/Cookie:Properties/Name/Description",
+					}
+				},
+			},
+		},
+		Constructors: rbxmk.Constructors{
+			"new": rbxmk.Constructor{
+				Func: func(s rbxmk.State) int {
+					name := string(s.Pull(1, "string").(types.String))
+					value := string(s.Pull(2, "string").(types.String))
+					cookie := rtypes.Cookie{Cookie: &http.Cookie{Name: name, Value: value}}
+					return s.Push(cookie)
+				},
+				Dump: func() dump.MultiFunction {
+					return dump.MultiFunction{
+						dump.Function{
+							Parameters: dump.Parameters{
+								{Name: "name", Type: dt.Prim("string")},
+								{Name: "value", Type: dt.Prim("string")},
+							},
+							Returns: dump.Parameters{
+								{Type: dt.Prim("Cookie")},
+							},
+							Summary:     "Types/Cookie:Constructors/new/Summary",
+							Description: "Types/Cookie:Constructors/new/Description",
+						},
 					}
 				},
 			},